gps: fix error when building with clang
* error: constant expression evaluates to 4294967295 which cannot be narrowed to type 'long'

+++ gps/utils/loc_log.h
@@ -41,7 +41,7 @@ extern "C"
 typedef struct
 {
    char                 name[128];
-   long                 val;
+   long long            val;
 } loc_name_val_s_type;
 
 #define NAME_VAL(x) {"" #x "", x }
